+++ test/blackbox-tests/test-cases/c-flags.t/run.t
@@ -12,7 +12,7 @@ bar.c is built with an "empty" set of flags.
   $ O_CCPPF=${O_CCPPF%% }
 
 
-future_c_and_cxx_flags_handling = default (false)
+use_standard_c_and_cxx_flags = default (false)
 ==================================
 
   $ cat >dune-project <<EOF
@@ -25,7 +25,7 @@ future_c_and_cxx_flags_handling = default (false)
   Warning: The flag set for these foreign sources overrides the `:standard` set
   of flags. However the flags in this standard set are still added to the
   compiler arguments by Dune. This might cause unexpected issues. You can
-  disable this warning by defining the option `(future_c_and_cxx_flags_handling
+  disable this warning by defining the option `(use_standard_c_and_cxx_flags
   <bool>)` in your `dune-project` file. Setting this option to `true` will
   effectively prevent Dune from silently adding c-flags to the compiler
   arguments which is the new recommended behaviour.
@@ -37,7 +37,7 @@ future_c_and_cxx_flags_handling = default (false)
   Warning: The flag set for these foreign sources overrides the `:standard` set
   of flags. However the flags in this standard set are still added to the
   compiler arguments by Dune. This might cause unexpected issues. You can
-  disable this warning by defining the option `(future_c_and_cxx_flags_handling
+  disable this warning by defining the option `(use_standard_c_and_cxx_flags
   <bool>)` in your `dune-project` file. Setting this option to `true` will
   effectively prevent Dune from silently adding c-flags to the compiler
   arguments which is the new recommended behaviour.
@@ -59,12 +59,12 @@ ocamlc_cpp flags appear in the compiler command line:
   $ cat out_bar | grep -ce "${O_CCF} ${O_CCPPF}"
   1
 
-future_c_and_cxx_flags_handling = true
+use_standard_c_and_cxx_flags = true
 =================================
 
   $ cat >dune-project <<EOF
   > (lang dune 2.8)
-  > (future_c_and_cxx_flags_handling true)
+  > (use_standard_c_and_cxx_flags true)
 
   $ dune rules -m foo.o | tr -s '\t\n\\' ' ' > out_foo
   $ dune rules -m bar.o | tr -s '\t\n\\' ' ' > out_bar
@@ -88,19 +88,19 @@ ocamlc_cpp are effectively removed from the compiler command line
   0
   [1]
 
-future_c_and_cxx_flags_handling = true but dune < 2.8
+use_standard_c_and_cxx_flags = true but dune < 2.8
 ================================================
 
   $ cat >dune-project <<EOF
   > (lang dune 2.7)
-  > (future_c_and_cxx_flags_handling true)
+  > (use_standard_c_and_cxx_flags true)
 
   $ dune rules
-  File "dune-project", line 2, characters 0-38:
-  2 | (future_c_and_cxx_flags_handling true)
-      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-  Error: 'future_c_and_cxx_flags_handling' is only available since version 2.8
-  of the dune language. Please update your dune-project file to have (lang dune
+  File "dune-project", line 2, characters 0-35:
+  2 | (use_standard_c_and_cxx_flags true)
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+  Error: 'use_standard_c_and_cxx_flags' is only available since version 2.8 of
+  the dune language. Please update your dune-project file to have (lang dune
   2.8).
   [1]